# docs/swagger.yaml
openapi: 3.0.2
info:
    title: Games-api
    version: 1.0.0 
    description: '### This is a Swagger UI for our Games API'
paths:
    '/games':
        get:
          tags:
              - games
          responses:
              '200':
                  content:
                      application/json:
                          schema:
                              type: array
                              items:
                                  type: string
                          examples:
                              List of games:
                                  value:
                                      - Wither 3
                                      - Cyberpunk 2077
                  description: Returns a list of games
          summary: Get a list of all games
          description: Returns a list of all games in API
        post:
          requestBody:
              description: A new game to be created
              content:
                  application/json:
                      schema:
                          $ref: '#/components/schemas/error'
                      examples:
                          Error:
                              value: Missing parameters
              required: true
          tags:
              - new-game
          responses:
              '201':
                  description: Game creating
              '400':
                  description: One or all params are missing
          summary: Creates a game
          description: Creates a new game

    '/games/{id}':
        get:
            tags:
                - details
            responses:
                '200':
                    content:
                        application/json:
                            examples:
                                Full details of a given game:
                                    value:
                                        id: 1
                                        name: Wither 3
                                        price: 30
                    description: Returns full details of a given id game
                '404':
                    content:
                        application/json:
                            examples:
                                Game /w+ was not found:
                                    value: Game \\w+ was not found in game array
                    description: Game was not found in games array
            operationId: id
            summary: Get full details of a game
            description: Returns full details of a given id game
        delete:
          tags:
              - delete
          parameters:
              -
                  name: id
                  schema:
                      type: string
                  in: path
                  required: true
          responses:
              '204':
                  content:
                      application/json:
                          schema:
                              $ref: '#/components/schemas/delete-game'
                  description: Game was deleted
              '404':
                  description: ID was not found
          summary: Deleting game by id
          description: wow
          parameters:
              -
                  name: id
                  in: path
                  required: true
components:
    schemas:
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message. 
# This is a Swagger UI for our Games API  \n\n### Error handling\nErrors are presented using the HTTP status codes 400-599. The response include a more detailed explanation about the error in the body.\n\n```\n{\n  \"error\": \"Cleartext explanation of the error that occurred\"\n}\n```\n\n### Generic API errors\nThese error responses can be returned by any of the described endpoints.\n\n| HTTP status  |  Description |\n|---|---|\n| 400| Bad input parameters or invalid request. More information can be found in the response body.\n| 401| The consumer does not have access to the requested resource or the provided token is not valid.\n| 404| The requested resource can not be found or the user does not have access to the resource.\n| 5XX| A server error has occurred while processing the request.